📱 Mobile Header, Footer, Form
Todo: how to get to resources page from mobile header?
