+++ src/app/globals.css
@@ -9,6 +9,7 @@
   --main-2: #E3B971;
   --main-2-b: #d9a03d;
   --main-3: #CD5A48;
+  --main-3-b: #ec7c6b;
   --main-4: #AD442C;
   --main-5: #803124;
   --main-6: #533624;
